requirements.txt
beautifulsoup4==4.6.0
bleach==2.1.1
certifi==2017.7.27.1
chardet==3.0.4
decorator==4.1.2
defusedxml==0.5.0
Django==1.11.6
django-allauth==0.31.0
django-disqus==0.5
django-el-pagination==3.1.0
django-endless-pagination==2.0
django-extensions==1.9.7
django-hitcount==1.2.3
django-pagination-bootstrap==1.2.4
entrypoints==0.2.3
gistory==0.42
html5lib==1.0b10
idna==2.6
ipykernel==4.6.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.3
jedi==0.11.0
Jinja2==2.9.6
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-core==4.4.0
MarkupSafe==1.0
mistune==0.8
mock==2.0.0
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.2.0
oauthlib==2.0.6
pandocfilters==1.4.2
parso==0.1.0
pbr==3.1.1
pexpect==4.2.1
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.3
pyzmq==16.0.3
requests==2.18.4
requests-oauthlib==0.8.0
simplegeneric==0.8.1
six==1.11.0
terminado==0.6
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
typing==3.6.2
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.0.6
